Fixes #195, ISBNs should not become INTs
The ISBNs were actually stored fine in the database--it was just that the DB methods were using getInt32() to retrieve them. Using getInt64() instead.

+++ chrome/chromeFiles/content/scholar/xpcom/db.js
@@ -456,7 +456,7 @@ Scholar.DB = new function(){
 		var type = statement.getTypeOfIndex(i);
 		switch (type){
 			case statement.VALUE_TYPE_INTEGER:
-				var func = statement.getInt32;
+				var func = statement.getInt64;
 				break;
 			case statement.VALUE_TYPE_TEXT:
 				var func = statement.getUTF8String;
